+++ pupa/tests/scrape/test_people_org_scrape.py
@@ -93,49 +93,50 @@ def test_org_add_post():
 
 
 def test_legislator_related_district():
-    l = Person('John Adams', district='1', primary_org='legislature')
-    l.pre_save('jurisdiction-id')
+    leg = Person('John Adams', district='1', primary_org='legislature')
+    leg.pre_save('jurisdiction-id')
 
-    assert len(l._related) == 1
-    assert l._related[0].person_id == l._id
-    assert get_pseudo_id(l._related[0].organization_id) == {'classification': 'legislature'}
-    assert get_pseudo_id(l._related[0].post_id) == {"organization__classification": "legislature",
-                                                    "label": "1"}
+    assert len(leg._related) == 1
+    assert leg._related[0].person_id == leg._id
+    assert get_pseudo_id(leg._related[0].organization_id) == {'classification': 'legislature'}
+    assert get_pseudo_id(leg._related[0].post_id) ==\
+        {"organization__classification": "legislature",
+         "label": "1"}
 
 
 def test_legislator_related_chamber_district():
-    l = Person('John Adams', district='1', primary_org='upper')
-    l.pre_save('jurisdiction-id')
+    leg = Person('John Adams', district='1', primary_org='upper')
+    leg.pre_save('jurisdiction-id')
 
-    assert len(l._related) == 1
-    assert l._related[0].person_id == l._id
-    assert get_pseudo_id(l._related[0].organization_id) == {'classification': 'upper'}
-    assert get_pseudo_id(l._related[0].post_id) == {"organization__classification": "upper",
-                                                    "label": "1"}
+    assert len(leg._related) == 1
+    assert leg._related[0].person_id == leg._id
+    assert get_pseudo_id(leg._related[0].organization_id) == {'classification': 'upper'}
+    assert get_pseudo_id(leg._related[0].post_id) == {"organization__classification": "upper",
+                                                      "label": "1"}
 
 
 def test_legislator_related_chamber_district_role():
-    l = Person('John Adams', district='1', primary_org='lower', role='Speaker')
-    l.pre_save('jurisdiction-id')
+    leg = Person('John Adams', district='1', primary_org='lower', role='Speaker')
+    leg.pre_save('jurisdiction-id')
 
-    assert len(l._related) == 1
-    assert l._related[0].person_id == l._id
-    assert get_pseudo_id(l._related[0].organization_id) == {'classification': 'lower'}
-    assert get_pseudo_id(l._related[0].post_id) == {"organization__classification": "lower",
-                                                    "label": "1", "role": "Speaker"}
-    assert l._related[0].role == 'Speaker'
+    assert len(leg._related) == 1
+    assert leg._related[0].person_id == leg._id
+    assert get_pseudo_id(leg._related[0].organization_id) == {'classification': 'lower'}
+    assert get_pseudo_id(leg._related[0].post_id) == {"organization__classification": "lower",
+                                                      "label": "1", "role": "Speaker"}
+    assert leg._related[0].role == 'Speaker'
 
 
 def test_legislator_related_party():
-    l = Person('John Adams', party='Democratic-Republican')
-    l.pre_save('jurisdiction-id')
+    leg = Person('John Adams', party='Democratic-Republican')
+    leg.pre_save('jurisdiction-id')
 
     # a party membership
-    assert len(l._related) == 1
-    assert l._related[0].person_id == l._id
-    assert get_pseudo_id(l._related[0].organization_id) == {'classification': 'party',
-                                                            'name': 'Democratic-Republican'}
-    assert l._related[0].role == 'member'
+    assert len(leg._related) == 1
+    assert leg._related[0].person_id == leg._id
+    assert get_pseudo_id(leg._related[0].organization_id) == {'classification': 'party',
+                                                              'name': 'Democratic-Republican'}
+    assert leg._related[0].role == 'member'
 
 
 def test_committee_add_member_person():